src/CollibraScripts/requirements.txt
requests
pandas
openpyxl
python-dotenv
jsonschema
